Match UIListHighlight and UIListLabel, more UI TUs added as well (#255)
* uicomponent and uiproxy ctor tweaks

* uilisthighlight work

* match uilisthighlight

* uilistlabel work

* uilistlabel work

* add uilistmesh file

* uilistmesh work

* more uilistmesh work

* uiliststate work

* nearly match uilistlabel

* almost

* uilistlabel ready to link

* match uilistlabel

* uilistmesh draw

* nearly match uilistmeshelement

* uilistmeshelement regswap

* uilistslot work

* uilistprovider work

* dataprovider work

---------

+++ src/system/ui/UIList.h
@@ -34,7 +34,7 @@ public:
     virtual void Update();
     virtual void AdjustTrans(Transform&, const UIListElementDrawState&){}
     virtual void AdjustTransSelected(Transform&){}
-    virtual short NumData() const;
+    virtual int NumData() const;
     virtual void StartScroll(const UIListState&, int, bool);
     virtual void CompleteScroll(const UIListState&);
     virtual int SelectedAux() const;
